Add separate messages folder and update Dockerfiles
