+++ messages/__init__.py
@@ -1,0 +1,3 @@
+"""Add submodules to the python path."""
+
+import init
