+++ station_component_manifest.yml
@@ -1,6 +1,6 @@
 # The component manifest file for the Station component
 
-Name: StationComponentTest
+Name: StationComponent
 Type: platform
 Description: "Station Component Test v1.0.1"
 DockerImage: station_component_v2
