+++ stationcomponent.Dockerfile
@@ -8,6 +8,7 @@ RUN mkdir -p /station_component
 RUN mkdir -p /init
 RUN mkdir -p /logs
 RUN mkdir -p /simulation-tools
+RUN mkdir -p /messages
 
 # install the python libraries inside the Docker image
 COPY requirements.txt /requirements.txt
@@ -16,6 +17,7 @@ RUN pip install -r /requirements.txt
 
 # copy the required directories with their content to the Docker image
 COPY station_component/ /station_component/
+COPY messages/ /messages/
 COPY init/ /init/
 COPY simulation-tools/ /simulation-tools/
