--- user_component/PowerOutput_message.py
@@ -1,105 +1,0 @@
-from __future__ import annotations
-from typing import Any, Dict, Optional
-
-from tools.exceptions.messages import MessageError, MessageValueError
-from tools.messages import AbstractResultMessage
-
-
-class PowerOutputMessage(AbstractResultMessage):
-    """Description for the SimpleMessage class"""
-
-    CLASS_MESSAGE_TYPE = "PowerOutput"
-    MESSAGE_TYPE_CHECK = True
-
-    POWER_OUTPUT_ATTRIBUTE = "PowerOutput"
-    POWER_OUTPUT_PROPERTY = "power_output"
-
-    STATION_ID_ATTRIBUTE = "StationID"
-    STATION_ID_PROPERTY = "station_id"
-
-    # all attributes specific that are added to the AbstractResult should be introduced here
-    MESSAGE_ATTRIBUTES = {
-        POWER_OUTPUT_ATTRIBUTE: POWER_OUTPUT_PROPERTY,
-        STATION_ID_ATTRIBUTE: STATION_ID_PROPERTY
-    }
-    # list all attributes that are optional here (use the JSON attribute names)
-    OPTIONAL_ATTRIBUTES = []
-
-    # all attributes that are using the Quantity block format should be listed here
-    QUANTITY_BLOCK_ATTRIBUTES = {}
-
-    # all attributes that are using the Quantity array block format should be listed here
-    QUANTITY_ARRAY_BLOCK_ATTRIBUTES = {}
-
-    # all attributes that are using the Time series block format should be listed here
-    TIMESERIES_BLOCK_ATTRIBUTES = []
-
-    # always include these definitions to update the full list of attributes to these class variables
-    # no need to modify anything here
-    MESSAGE_ATTRIBUTES_FULL = {
-        **AbstractResultMessage.MESSAGE_ATTRIBUTES_FULL,
-        **MESSAGE_ATTRIBUTES
-    }
-    OPTIONAL_ATTRIBUTES_FULL = AbstractResultMessage.OPTIONAL_ATTRIBUTES_FULL + OPTIONAL_ATTRIBUTES
-    QUANTITY_BLOCK_ATTRIBUTES_FULL = {
-        **AbstractResultMessage.QUANTITY_BLOCK_ATTRIBUTES_FULL,
-        **QUANTITY_BLOCK_ATTRIBUTES
-    }
-    QUANTITY_ARRAY_BLOCK_ATTRIBUTES_FULL = {
-        **AbstractResultMessage.QUANTITY_ARRAY_BLOCK_ATTRIBUTES_FULL,
-        **QUANTITY_ARRAY_BLOCK_ATTRIBUTES
-    }
-    TIMESERIES_BLOCK_ATTRIBUTES_FULL = (
-        AbstractResultMessage.TIMESERIES_BLOCK_ATTRIBUTES_FULL +
-        TIMESERIES_BLOCK_ATTRIBUTES
-    )
-
-    # for each attributes added by this message type provide a property function to get the value of the attribute
-    # the name of the properties must correspond to the names given in MESSAGE_ATTRIBUTES
-    # template for one property:
-
-
-    @property
-    def power_output(self) -> int:
-        return self.__power_output
-
-    @property
-    def station_id(self) -> str:
-        return self.__station_id
-
-
-    @power_output.setter
-    def power_output(self, power_output: int):
-        self.__power_output = power_output
-
-    @station_id.setter
-    def station_id(self, station_id: str):
-        self.__station_id = station_id
-
-    def __eq__(self, other: Any) -> bool:
-        return (
-            super().__eq__(other) and
-            isinstance(other, PowerOutputMessage) and
-            self.station_id == other.station_id and
-            self.power_output == other.power_output
-        )
-
-    @classmethod
-    def _check_station_id(cls, power_output: str) -> bool:
-        return isinstance(power_output, str)
-
-    @classmethod
-    def _check_power_output(cls, power_output: int) -> bool:
-        return isinstance(power_output, int)
-
-    @classmethod
-    def from_json(cls, json_message: Dict[str, Any]) -> Optional[PowerOutputMessage]:
-        """TODO: description for the from_json method"""
-        try:
-            message_object = cls(**json_message)
-            return message_object
-        except (TypeError, ValueError, MessageError):
-            return None
-
-
-PowerOutputMessage.register_to_factory()
